requirements.txt
PySide6>=6.5.0
cryptography>=41.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
keyboard>=0.13.5          # 用于全局快捷键（Windows/Linux）